+++ docs/examples/snek5000-tgv/src/snek5000_tgv/output.py
@@ -1,10 +1,14 @@
-from snek5000_tgv.templates import box, makefile_usr, size
-
-from snek5000 import mpi
 from snek5000.output.base import Output as OutputBase
+from snek5000.resources import get_base_templates
+
+box, makefile_usr, size = get_base_templates()
 
 
 class OutputTGV(OutputBase):
+    template_box = box
+    template_makefile_usr = makefile_usr
+    template_size = size
+
     @property
     def makefile_usr_sources(self):
         """
@@ -12,15 +16,6 @@ class OutputTGV(OutputBase):
         Dict[directory]  -> list of source files
         """
         return {}
-
-    def post_init(self):
-        super().post_init()
-
-        # Write additional source files to compile the simulation
-        if mpi.rank == 0 and self._has_to_save and self.sim.params.NEW_DIR_RESULTS:
-            self.write_box(box)
-            self.write_size(size)
-            self.write_makefile_usr(makefile_usr)
 
     @classmethod
     def _set_info_solver_classes(cls, classes):