+++ src/snek5000/output/base.py
@@ -769,6 +769,18 @@ class Output(OutputCore):
             self.copy(self.path_run)
             config = self.write_snakemake_config()
             self.build_nek5000(config)
+            self.post_init_create_additional_source_files()
+
+    def post_init_create_additional_source_files(self):
+        """Create the .box, SIZE and makefile_usr files from their template"""
+        for name in ("box", "size", "makefile_usr"):
+            try:
+                template = getattr(self, f"template_{name}")
+            except AttributeError:
+                pass
+            else:
+                if template is not None:
+                    getattr(self, f"write_{name}")(template)
 
     def _save_info_solver_params_xml(self, replace=False):
         """Saves the par file, along with ``params_simul.xml`` and